+++ repology_client/constants.py
@@ -9,7 +9,7 @@ Hardcoded constants for Repology API.
 PACKAGE = "repology-client"
 
 #: Library version.
-VERSION = "0.4.0"
+VERSION = "0.5.0"
 
 #: Library homepage.
 HOMEPAGE = "https://repology-client.sysrq.in"